pyproject.toml
[project]
name = "media-workflow"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "aioboto3>=13.3.0",
    "aiohttp>=3.11.11",
    "cairosvg>=2.7.1",
    "fonttools>=4.55.3",
    "imageio[pillow-heif,tifffile]>=2.36.1",
    "jinja2>=3.1.5",
    "json-repair>=0.32.0",
    "numpy>=2.2.1",
    "openai>=1.58.1",
    "opencv-python>=4.10.0.84",
    "opentelemetry-exporter-otlp-proto-http>=1.29.0",
    "pedantic>=2.1.8",
    "pillow>=11.0.0",
    "psd-tools>=1.10.4",
    "pymupdf>=1.25.1",
    "scikit-learn>=1.6.0",
    "temporalio[opentelemetry]>=1.8.0",
]

[dependency-groups]
dev = [
    "ipython>=8.31.0",
    "pre-commit>=4.0.1",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.0",
    "pytest-xdist>=3.6.1",
    "ruff>=0.9.1",
]
c4d = [
    "aioboto3>=13.3.0",
    "aiohttp>=3.11.11",
    "opentelemetry-exporter-otlp-proto-http>=1.29.0",
    "temporalio[opentelemetry]>=1.8.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
filterwarnings = ["ignore"]

[tool.ruff]
lint.extend-select = ["I", "E", "W", "N", "D", "S", "B", "A", "C4"]
lint.ignore = ["D10", "S101", "D401"]
lint.pydocstyle.convention = "pep257"
line-length = 120
extend-exclude = ["pylette"]

[tool.pyright]
include = ["media_workflow"]
pythonVersion = "3.13"
